requirements.txt
addict==2.4.0
clip==0.2.0
einops==0.8.0
flash_attn==2.6.3
imageio==2.34.1
laspy==2.5.4
matplotlib==3.5.3
MinkowskiEngine==0.5.4
mmcv==2.2.0
numpy==1.24.4
ocnn==2.2.2
open3d==0.13.0
opencv_python==4.9.0.80
pandas==2.0.3
Pillow==10.4.0
pyquaternion==0.9.9
PyYAML==6.0.1
PyYAML==6.0.2
raillabel==3.1.0
regex==2024.9.11
scipy==1.14.1
seaborn==0.13.2
setuptools==72.1.0
setuptools==45.2.0
SharedArray==3.2.4
SharedArray==3.2.4
spconv_cu120==2.3.6
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
termcolor==2.4.0
timm==1.0.9
torch==2.3.0
torch_geometric==2.6.0
torch_points3d==1.3.0
torch_points_kernels==0.7.0
torch_scatter==2.1.2+pt23cu121
tqdm==4.66.2
trimesh==4.4.9
wandb==0.18.1
waymo_open_dataset==1.0.1
yapf==0.40.1
